% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bll_flow.R
\name{bll_flow}
\alias{bll_flow}
\title{Makes an object an instance of bll_flow and returns it. This function should
be run on all the variables in the initial data set}
\usage{
bll_flow(x, label = "Not Defined")
}
\arguments{
\item{x}{numeric The object to convert to bllFlow}

\item{label}{string Optional label that will be attached to the object}
}
\value{
Returns a new bll_flow object
}
\description{
Makes an object an instance of bll_flow and returns it. This function should
be run on all the variables in the initial data set
}
