% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{buildInteractionsList}
\alias{buildInteractionsList}
\title{Builds the interactions list for the provided rows}
\usage{
buildInteractionsList(rowsWithInteractions)
}
\arguments{
\item{rowsWithInteractions}{A webspec dataframe that contains only
interactions rows}
}
\value{
list[list] A list where each item is a 2 item list containing the variables
being interacted
}
\description{
Builds the interactions list for the provided rows
}
