% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-builder.R
\name{buildProjectFiles}
\alias{buildProjectFiles}
\title{Builds several R files from the web spec csv file located at the path
specified by the webSpecCsvFilePath argument. The R files are written to the folder whose path
is specified in the projectFolderPath arg}
\usage{
buildProjectFiles(webSpecCsvFilePath, projectFolderPath)
}
\arguments{
\item{webSpecCsvFilePath}{The path to the web spec csv file}

\item{projectFolderPath}{The path to the folder where the R files will be put}
}
\description{
Builds several R files from the web spec csv file located at the path
specified by the webSpecCsvFilePath argument. The R files are written to the folder whose path
is specified in the projectFolderPath arg
}
