% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationPlot.R
\name{CalibrationPlot}
\alias{CalibrationPlot}
\title{Calibration Plot
Creates a calibration plot using plotly package. Returns an actual vs predicted plot. Used for model fit visualization}
\usage{
CalibrationPlot(data, observed, predicted, title)
}
\arguments{
\item{data}{a data.frame}

\item{observed}{variable of observed observation from data.frame, object of class string}

\item{predicted}{variable of predicted observations from data.frame, object of class string}

\item{title}{an overall title for the plot (use NA for no title), object of class string}
}
\description{
Calibration Plot
Creates a calibration plot using plotly package. Returns an actual vs predicted plot. Used for model fit visualization
}
\examples{
##Load packages
library(plotly)

#Load Female risk decile data from Mortality Population Risk csv
data <- as.data.frame(read.csv("inst/extdata/RiskDecile.csv")) 

##check variables for appropriate observed and predicted variable names
head(data, 2L)

##View plot - will retrun plot generated in plotly in R studio viewer
CalibrationPlot(data,'observed','predicted', 'Female Risk Decile')  
}
\author{
Molly Campbell
}
