% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Generic center function. Should not be modified}
\usage{
center(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Generic center function. Should not be modified
}
