% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.Max}
\alias{clean.Max}
\title{Generic function for clean.Max}
\usage{
clean.Max(bllFlowModel, ...)
}
\arguments{
\item{bllFlowModel}{The bllFlowModel we will clean}

\item{...}{Arguments to the next method in the chain}
}
\description{
Generic function for clean.Max
}
