% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.Min}
\alias{clean.Min}
\title{Generic function for clean.Min}
\usage{
clean.Min(bllFlowModel, ...)
}
\arguments{
\item{bllFlowModel}{The bllflow model we will clean}

\item{...}{Arguments to the next method in the chain}
}
\description{
Generic function for clean.Min
}
