% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMessyData.R
\name{dirtyData}
\alias{dirtyData}
\title{Makes 'messy' data
makes invalid and missing data to a dataset that has been previously 
cleaned.}
\usage{
dirtyData(datasetname, variablename, messy0missing1, prc, outdataset)
}
\arguments{
\item{datasetname}{The name of the original dataset that has the variable 
to be messed up}

\item{variablename}{The name of the original variable to be messed up}

\item{messy0missing1}{value = 0 to introduce messy data, 1 to introduce 
missing values}

\item{prc}{of messy data/missing values to be introduced. 
e.g. 0.05 for 5 percent messy/missing}

\item{outdataset}{Dataset to have original and messed up data. 
Messed up variable has the same name with "dirty" prefix}
}
\value{
dataframe with 'messy' data added
}
\description{
Use cases include:
 - teaching how to clean a dataset or impute missing data
 - unit testing cleaning and data imputation functons
 - unit testing predictive algoirthm scoring engines.
}
\examples{
These are more like 'vingettes' TODO: make a vingett and change
to simple function examples.
originaldata<-as.data.frame(read.csv('messyd.csv'))
dirtyData(originaldata,BMI, 1, 0.05, dirtyData)

sink("mess")
dirtydata
 sink()

dirtyData(originaldata,BMI, 0, 0.05, missiongData)
sink("miss")
missingdata
sink()
}
