% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{formatInteractionVariable}
\alias{formatInteractionVariable}
\title{Formats a variable name for interaction}
\usage{
formatInteractionVariable(varName)
}
\arguments{
\item{varName}{The var name to format. This is the not the entire interaction
var string but just one of the variables part of the interaction}
}
\value{
string
}
\description{
Formats a variable name for interaction
}
