% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered-variable.R
\name{getCodeForCenteredVariable}
\alias{getCodeForCenteredVariable}
\title{Returns R code to center a variable}
\usage{
getCodeForCenteredVariable(variableNameToCenter)
}
\arguments{
\item{variableNameToCenter}{The name of the variable to center}
}
\value{
string
}
\description{
Returns R code to center a variable
}
