% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-variable.R
\name{getCodeForDummyVariable}
\alias{getCodeForDummyVariable}
\title{Returns variable which contains dummying code for a categorical variable}
\usage{
getCodeForDummyVariable(dummyVariableName, numberOfCategories)
}
\arguments{
\item{dummyVariableName}{The name of the categorical variable}

\item{numberOfCategories}{The number of categories for the variable}
}
\value{
string
}
\description{
Returns variable which contains dummying code for a categorical variable
}
\examples{

}
