% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{getCodeForInteraction}
\alias{getCodeForInteraction}
\title{Returns code for an interaction}
\usage{
getCodeForInteraction(varOne, varTwo)
}
\arguments{
\item{varOne}{The first var part of this interaction}

\item{varTwo}{Th second var part of this interaction}
}
\value{
string
}
\description{
Returns code for an interaction
}
