% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{isInInteractionsList}
\alias{isInInteractionsList}
\title{Checks whether an interactions is already present in the interactionsList arg}
\usage{
isInInteractionsList(interactionsList, varOne, varTwo)
}
\arguments{
\item{interactionsList}{list[list[string, string]] The interactions to check}

\item{varOne}{One of the variables part of the interaction}

\item{varTwo}{The second var part of the interaction}
}
\value{
boolean
}
\description{
Checks whether an interactions is already present in the interactionsList arg
}
