% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-builder.R
\name{isPortWebSpecFile}
\alias{isPortWebSpecFile}
\title{Checks whether a web spec file is for a PORT algorithm or for a RESPECT
algorithm}
\usage{
isPortWebSpecFile(webSpecs)
}
\arguments{
\item{webSpecs}{The web spec file rows in a data frame}
}
\value{
boolean
}
\description{
Checks whether a web spec file is for a PORT algorithm or for a RESPECT
algorithm
}
