% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-function-activity.R
\name{LogFunctionActivity}
\alias{LogFunctionActivity}
\title{Title}
\usage{
LogFunctionActivity(bllFlow, rowsChecked, rowsAffected, actionTaken,
  reason, executedFunction, variable, value)
}
\arguments{
\item{bllFlow}{}
}
\value{

}
\description{
Title
}
