% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-date.R, R/make-event-date.R
\name{make_event_date}
\alias{make_event_date}
\title{Add an event date, when you only have an event probability}
\usage{
make_event_date(df, event_probability, followup_time,
  label = "Time to event", units = NA)

make_event_date(df, event_probability, followup_time,
  label = "Time to event", units = NA)
}
\arguments{
\item{df}{(required) Data frame with event propobablity. Each row represents
an observation.}

\item{event_probability}{(required) Target event probablity. Event probability.
e.g. 5-year risk of death is 0.2 represents 20 \% proabality of dying over
5 years or 1825 days.}

\item{followup_time}{(required) Time period for event probablity.}

\item{label}{(optional) Name of event. If missing, default to 'Time to
event'.}

\item{units}{(optional) Time unit for event. e.g. days.
#' @return List with event times. Returned list length = list length of df. Event
date =  totalTime (censored), if no event.}

\item{df}{(required) Data frame with event propobablity. Each row represents
an observation.}

\item{event_probability}{(required) Target event probablity. Event probability.
e.g. 5-year risk of death is 0.2 represents 20 \% proabality of dying over
5 years or 1825 days.}

\item{followup_time}{(required) Time period for event probablity.}

\item{label}{(optional) Name of event. If missing, default to 'Time to
event'.}

\item{units}{(optional) Time unit for event. e.g. days.
#' @return List with event times. Returned list length = list length of df. Event
date =  totalTime (censored), if no event.}
}
\description{
To create example data for algorithm development.

Use case: You have validation scoring data with event probablity but not
event date or censoring date.

Use case: You have cohort data with only exposure variables and you would
like to create a fictious event or censoring events.

To create example data for algorithm development.

Use case: You have validation scoring data with event probablity but not
event date or censoring date.

Use case: You have cohort data with only exposure variables and you would
like to create a fictious event or censoring events.
}
\examples{
# load the test data in the bllFlow package
\dontrun{
 df <- as.data.frame(read.csv("inst/extdata/RESPECT-EOL_validation.csv"))
 
 # or 
 # df <- RESPECT-EOL_validation
 
# each ovservation as event probablity of df$risk
  df$tt_event <- make_event_date(df, df$risk, 1825, units = 'days')
  
 # add variable labeles
   
  
# check the label
  label(df$tt_event)
 
# check to see if events have been added
  table(df$tt_event)
 
# check to see if df$tt_event is about the same as df$risk
  mean(df$risk)
  nrow(df[df$tt_event < 1825,]) / nrow(df)

# make other event data like censoring or competing events
  df$ttcensor <- make_event_date(df, .01, 1825)
}
# load the test data in the bllFlow package
\dontrun{
 df <- as.data.frame(read.csv("inst/extdata/RESPECT-EOL_validation.csv"))
 
 # or 
 # df <- RESPECT-EOL_validation
 
# each ovservation as event probablity of df$risk
  df$tt_event <- make_event_date(df, df$risk, 1825, units = 'days')
  
 # add variable labeles
   
  
# check the label
  label(df$tt_event)
 
# check to see if events have been added
  table(df$tt_event)
 
# check to see if df$tt_event is about the same as df$risk
  mean(df$risk)
  nrow(df[df$tt_event < 1825,]) / nrow(df)

# make other event data like censoring or competing events
  df$ttcensor <- make_event_date(df, .01, 1825)
}
}
\keyword{make_data}
