% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-date.R
\name{makeEventDate}
\alias{makeEventDate}
\title{Add an event date, when you only have an event probability}
\usage{
makeEventDate(df, eventProbability, followUpTime,
  label = "Time to event")
}
\arguments{
\item{df}{(required) Data frame with event propobablity. Each row represents
an observation.}

\item{eventProbability}{(required) Target event probablity. Event probability
e.g. 5-year risk of death is 0.2 represents 20 \% proabality of dying over
5 years or 1825 days.}

\item{followUpTime}{(required) Time period for event probablity}

\item{label}{Internal function from Hmisc used for adding labels.
#' @return List with event times. Returned list length = list length of df. Event
date =  totalTime (censored), if no event.}

\item{eventLabel}{(optional) Name of event. If missing, default to 'Time to
event'.}

\item{eventUnit}{(optional) Time unit for event. e.g. days}

\item{probablityWithdraw}{(optional) Probablity the observation was lost to
follow-up prior to end of follow up (followUpTime)}

\item{probablityCompetingEvent}{(optional) Probablity the observation was a
competing event.}
}
\description{
To create example data for algorithm development.

Use case: You have validation scoring data with event probablity but not
event date or censoring date.

Use case: You have cohort data with only exposure variables and you would
like to create a fictious event or censoring events.
}
\examples{
## Not run:
# df <- as.data.frame(read.csv("inst/extdata/RESPECT-EOL_validation.csv"))

## End(Not run)
# each ovservation as event probablity of df$risk
df$ttEvent <- makeEventDate(df, df$risk, 1825)
# add variable labeles
df$ttEvent <- makeEventDate(df, df$risk, 1825, eventLabel = "time to death (all-cause)")
#check the label
label(df$ttEvent)

df$ttcensor <- makeEventDate(df, .01, 1825)
#
}
\keyword{internal}
\keyword{label}
