% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP-plot.R
\name{op_plot}
\alias{op_plot}
\title{Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.}
\usage{
op_plot(data, observed, predicted, title=NA, xlab= NA)
}
\arguments{
\item{data}{a data.frame with observed events and predicted risk estimates}

\item{observed}{variable of observed observation from data.frame, object of class string}

\item{predicted}{variable of predicted observations from data.frame, object of class string}

\item{title}{an overall title for the plot (optional), object of class string}

\item{xlab}{a title for the x axis (optional), object of class string}
}
\description{
Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.
}
\examples{
#upload MPorT table one data set
mport <- read.csv("https://raw.githubusercontent.com/Big-Life-Lab/bllFlow/data-visualization/inst/extdata/MPoRT-table-one.csv?token=ApmTWVvRI4hGeBRlPLR6R3P6J73Po5Paks5cWzfswA\%3D\%3D", 
   fileEncoding="UTF-8-BOM")

#create character vector of columns needed for visualizations
col <- c("observed_risk_1_year", "predicted_risk_1_year", "group_by_sex", "group_by_1", "group_by_2", "group_by_value_label_1")

#create new data frame using filter_table_1 helper function 
data<-filter_table_1(mport, col, groupby_sex="Male", groupby_1 ="Drinks per week")

op_plot(data, "observed_risk_1_year", "predicted_risk_1_year", title="Observed vs Predicted for Male Drinks per Week" , xlab="Drinks per Week")
}
\author{
Molly Campbell
}
