% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPlot.R
\name{OvP}
\alias{OvP}
\title{Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.}
\usage{
OvP(data, observed, predicted, title, ylab)
}
\arguments{
\item{data}{a data.frame}

\item{observed}{variable of observed observation from data.frame, object of class string}

\item{predicted}{variable of predicted observations from data.frame, object of class string}

\item{title}{an overall title for the plot (use NA for no title), object of class string}

\item{ylab}{a title for the y axis (use NA for no title), object of class string}
}
\description{
Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.
}
\examples{

##Load packages
library(plotly)
library(curl)
##Load Female risk decile data from Mortality Population Risk csv from GitHub
data<-read.csv(curl("https://raw.githubusercontent.com/Big-Life-Lab/MPoRTv2/master/Development/Datasets/RiskDecile.csv\%5C?token=ApmTWZ2LBGKyfkp##PvTHUG9-f_w0D0Nb4ks5b5J1zwA\%3D\%3D"))

##view plot - will return plot generated in plotly in R studio viewer
OvP(data, 'observed', 'predicted', 'Observed vs Predicted for Female Risk Decile', ‘Risk Decile’)  
}
\author{
Molly Campbell
}
