% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPPlot.R
\name{OvPPlot}
\alias{OvPPlot}
\title{Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.}
\usage{
OvPPlot(data, observed, predicted, title, ylab)
}
\arguments{
\item{data}{a data.frame with observed events and predicted risk estimates}

\item{observed}{variable of observed observation from data.frame, object of class string}

\item{predicted}{variable of predicted observations from data.frame, object of class string}

\item{title}{an overall title for the plot (optional), object of class string}

\item{ylab}{a title for the y axis (optional), object of class string}
}
\description{
Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.
}
\examples{
##Load packages
library(plotly)
#Load Female risk decile data from Mortality Population Risk csv from GitHub
data <- data.frame(read.csv("inst/extdata/RiskDecile.csv")) 

#view plot - will return plot generated in plotly in R studio viewer
OvPPlot(data, 'observed', 'predicted', 'Observed vs Predicted for Female Risk Decile', 'Risk Decile')  
}
\author{
Molly Campbell
}
