% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-metadata-log.R
\name{print.metaDataLog}
\alias{print.metaDataLog}
\title{Custom print function for the bllFlow metaData log}
\usage{
\method{print}{metaDataLog}(x, ...)
}
\arguments{
\item{x}{The metaData log object}

\item{...}{Test}
}
\description{
Custom print function for the bllFlow metaData log
}
