% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bll-flow.R
\name{BLLFlow}
\alias{BLLFlow}
\title{Creates a bllflow model}
\usage{
BLLFlow(data = NULL, variables = NULL, variableDetails = NULL,
  ddi = NULL)
}
\arguments{
\item{data}{A dataframe that represents the dataset the model will be developed
on}

\item{variables}{A dataframe that has the specification sheet for this model. An example
of this worksheet is available here
\url{https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=0}.}

\item{variableDetails}{A dataframe that is the variable details worksheet. An example
of this worksheet is available here
\url{https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=1196358036}.}

\item{ddi}{An optinal string that contains the path to a ddi document}
}
\value{
A named list which is an instance of the bllflow class. The items
in the list are specified below: \cr
1. data - A dataframe that contains the passed data argument \cr
2. variables - A dataframe that contains the passed variables argument \cr
3. variableDetails - A dataframe that contains the passed variableDetails argument \cr
4. ddi - A string that contains the passed ddi argument \cr
5. additionalDDIMetaData - A named list. See the return type of the \code{\link{GetDDIDescription}} function \cr
6. populatedVariableDeatails - A dataframe that contains the rows in the variableDetails \cr
argument but with additional data filled in using the ddi argument it's specified
}
\description{
Wraps up the data, variables and variableDetails arguments in an R object, 
making it an instance of a bllflow class and returning the resulting object.
If a ddi argument is provided, all the metadata from the DDI document is 
imported into the R object
}
\examples{
# ALl the libraries we will be using
library(bllflow)
library(survival)

# Read in the data we will use for this example
data(pbc)

# Read in the variables and variable details CSV sheets which are part of the 
# master specification workbook
variablesSheet <- read.csv(system.file("extdata", "PBC-variables.csv", package="bllflow"))
variableDetails <- read.csv(system.file("extdata", "PBC-variableDetails.csv", package="bllflow"))

# Create a bllFlow R object for the PBC model using the above variables as args
# and store it in the pbcModel variable
pbcModel <- bllflow::BLLFlow(pbc, variablesSheet, variableDetails)

# The pbcModel variable is an R object of instance BLLFlow
print(attr(pbcModel, 'class'))
}
