% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bll_flow.R
\name{BLLFlow}
\alias{BLLFlow}
\title{Creates a BLLFlow instance from the data arg}
\usage{
BLLFlow(data, variables, variableDetailsSheet)
}
\arguments{
\item{data}{A dataframe that represents the dataset the model will be developed
on}

\item{variables}{The specification sheet for this model. An example
of this worksheet is available here
https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=0.
CSV file should be read in using the read.csv function.}

\item{variableDetailsSheet}{The variable details worksheet. An example
of this worksheet is available here
https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=1196358036.
CSV file should be read in using the read.csv function.}
}
\value{
Returns a new BllFlow object with the data, variables and
variableDetailsSheet args attached as variables.
}
\description{
Creates a BLLFlow instance from the data arg
}
\examples{
#' # Install the packages
 
# Use to generate the table one object
install.packages("tableone")
# Has the data we will use to generate a table one
install.packages("survival")

# Read in the data we will use to generate Table One

library(survival)
data(pbc)

# Read in the MSW and variable_details sheet for the PBC model
variablesSheet <- read.csv(file.path(getwd(), 'inst/extdata/PBC/PBC - variables.csv'))
variableDetailsSheet <- read.csv(file.path(getwd(), 'inst/extdata/PBC/PBC - variable_details.csv'))

# Create a bllFlow R object for the PBC model using the above variables as args
library(bllFlow)
pbcModel <- BLLFlow(pbc, variablesSheet, variableDetailsSheet)

# Passing objects other then a dataframe will create errors
pbcModel <- BLLFlow(pbc, c(1,2,3), list(2,3,4))
}
