% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bll_flow.R
\name{BLLFlow}
\alias{BLLFlow}
\title{Creates a BLLFlow instance from the data arg}
\usage{
BLLFlow(data, variables, variableDetailsSheet)
}
\arguments{
\item{data}{A dataframe that represents the dataset the model will be developed
on}

\item{variables}{The specification sheet for this model. An example
of this worksheet is available here
https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=0.
CSV file should be read in using the read.csv function.}

\item{variableDetailsSheet}{The variable details worksheet. An example
of this worksheet is available here
https://docs.google.com/spreadsheets/d/1QVqLKy_C185hzeQdJeOy-EeFMBXui1hZ1cB2sKqPG-4/edit#gid=1196358036.
CSV file should be read in using the read.csv function.
The paramater that is passed is the output of the read.csv function}
}
\value{
Returns the data argument with BLLFlow attached as a class name.
Adds the variables and variableDetailsSheet args as variables on the
data argument.
}
\description{
Creates a BLLFlow instance from the data arg
}
