% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Calibration Plot
Creates a calibration plot using plotly package. Returns an actual vs 
predicted plot. Used for model fit visualization}
\usage{
calibration_plot(data, observed, predicted, error=0, title=NA)
}
\arguments{
\item{data}{a data.frame}

\item{observed}{variable of observed observation from data.frame, object of 
class string}

\item{predicted}{variable of predicted observations from data.frame, object 
of class string}

\item{error}{percent error for error bars (optional)}

\item{title}{an overall title for the plot (optional), object of class string}
}
\description{
Calibration Plot
Creates a calibration plot using plotly package. Returns an actual vs 
predicted plot. Used for model fit visualization
}
\examples{
#upload MPorT table one data set
mport <- read.csv("https://raw.githubusercontent.com/Big-Life-Lab/bllFlow/data-visualization/inst/extdata/MPoRT-table-one.csv?token=ApmTWVvRI4hGeBRlPLR6R3P6J73Po5Paks5cWzfswA\%3D\%3D", 
fileEncoding="UTF-8-BOM")

#create character vector of columns needed for visualizations
col <- c("observed_risk_1_year", "predicted_risk_1_year", "group_by_sex", "group_by_1", "group_by_2", "group_by_value_label_1")

#create new data frame using filter_table_1 helper function 
data<-filter_table_1(mport, col, groupby_sex="Female", groupby_1="age")

calibration_plot(data, "observed_risk_1_year", "predicted_risk_1_year", title="Observed vs Predicted for Female age")
}
\author{
Molly Campbell
}
