% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{Calibration Plot}
\description
{Creates a calibration plot using plotly package. Returns an actual vs 
predicted plot. Used for model fit visualization}
\usage{
CalibrationPlot(data, observed, predicted, error=0, title=NA)
}
\arguments{
\item{data}{a data.frame}

\item{observed}{variable of observed observation from data.frame, object of 
class string}

\item{predicted}{variable of predicted observations from data.frame, object 
of class string}

\item{error}{percent error for error bars (optional)}

\item{title}{an overall title for the plot (optional), object of class string}
}
\description{
Calibration Plot
Creates a calibration plot using plotly package. Returns an actual vs 
predicted plot. Used for model fit visualization
}
\author{
Molly Campbell
}
