% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center.bll_flow}
\alias{center.bll_flow}
\title{Centers the x argument using the provided mean argument and returns a new bll_flow
variable}
\usage{
center.bll_flow(x, mean)
}
\arguments{
\item{x}{numeric,bll_flow The bll_flow variable to center}

\item{mean}{numeric The mean to center the x arg on}
}
\value{
numeric,bll_flow The centered object with attribute (attr) = center and 'label' = 'Centered'
}
\description{
Centers the x argument using the provided mean argument and returns a new bll_flow
variable
}
