% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.Max.BLLFlow}
\alias{clean.Max.BLLFlow}
\title{Cleans the data using the max and outlier columns in the variables sheet of
the MSW. Outlier method is applied on a row if any of the variable
values for that row is greater than the max value as specified in the variables
sheet. Outlier checking for the column is not applied if max value is NA.}
\usage{
\method{clean.Max}{BLLFlow}(bllFlowModel, print = FALSE, ...)
}
\arguments{
\item{bllFlowModel}{bllFlow model created using the BLLFlow method}

\item{print}{If set to true, prints the following metrics about the model:
1. Number of rows deleted from the dataset}

\item{...}{Arguments for next methods in the chain}
}
\value{
bllFlowModel that has had its data modified by the paramaters located in
the variables object
}
\description{
Currently supported outlier methods are:
Delete - Specified as 'delete' in MSW. Deletes the row from the data.
Deleted rows are stored in the metadata variable under the deletedRows name.
Missing - Specified as 'missing' in MSW. Column value for that row which does
not meet the criteria is set to NA.
Not Applicable - TODO.
Set to value - Specified as a number value in MSW. Column value for the row is
set to the value specified in the outlier column.
}
\examples{
# Install the packages

# Read in the data we will use

library(survival)
data(pbc)

# Read in the MSW and variable_details sheet for the PBC model
variablesSheet <- read.csv(file.path(getwd(),
 'bllFlow/extdata/PBC/PBC-variables.csv'))
variableDetailsSheet <- read.csv(file.path(getwd(),
 'bllFlow/extdata/PBC/PBC-variable-details.csv'))

# Create a bllFlow R object for the PBC model using the above variables as args
library(bllFlow)
pbcModel <- BLLFlow(pbc, variablesSheet, variableDetailsSheet)
# passing non dataframe objects when generating the bllFlow object will cause errors
#pbcModel <- BLLFlow(pbc, c(1,2,3), variableDetailsSheet)

# Clean the data
cleanedPbcModel <- clean.Max(pbcModel)

# if u wish to be updated in the log on what the function does set print to true
cleanedPbcModel <- clean.Max(cleanedPbcModel, print=TRUE)
}
