% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.Min}
\alias{clean.Min}
\title{Clean.Min}
\usage{
clean.Min(modelToClean, print = FALSE)
}
\arguments{
\item{modelToClean}{A passed object that contains the data that needs to be cleaned
as well as specifications on how to clean it.}

\item{print}{If set true prints human readable log of the performed cleaning operations}
}
\value{
The passed object with now cleaned data as well as a log containing the performed actions
}
\description{
Clean The data using the Min according to the specifications in the passed object.
}
