% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean.Min}
\alias{clean.Min}
\alias{clean.Min.BLLFlow}
\title{Clean a dataset by updating values below a certain minimum}
\usage{
clean.Min(bllFlowModel, ...)

\method{clean.Min}{BLLFlow}(bllFlowModel, print = FALSE, ...)
}
\arguments{
\item{bllFlowModel}{The bllflow model we will clean}

\item{...}{Arguments to the next method in the chain}

\item{print}{A boolean which when set to TRUE prints logs of what the operation did}
}
\value{
A bllflow named list whose dataset was cleaned
}
\description{
Clean a dataset by updating values below a certain minimum
}
\section{Methods (by class)}{
\itemize{
\item \code{BLLFlow}: Cleans the data using the min and outlier columns in the variables sheet of
the MSW. Outlier method is applied on a row if any of the variable
values for that row is less than the min value as specified in the variables
sheet. Outlier checking for the column is not applied if min value is NA.

Currently supported outlier methods are: \cr
1. \strong{Delete} - Specified as 'delete' in MSW. Deletes the row from the data. \cr
number of deleted rows as well as their reason for deletion is stored in the
metaData variable under the deletedRows name. \cr
2. \strong{Missing} - Specified as 'missing' in MSW. Column value for that row which does
not meet the criteria is set to NA. \cr
3. \strong{Not Applicable} - TODO. \cr
4. \strong{Set to value} - Specified as a number value in MSW. Column value for the row is
set to the value specified in the outlier column.
}}

\examples{
# Load packages
library(survival)
library(bllflow)

# Read in the data we will use
data(pbc)

# Read in the MSW and variable_details sheet for the PBC model
variablesSheet <- read.csv(system.file("extdata", "PBC-variables.csv", package="bllflow"))
variableDetailsSheet <- read.csv(system.file("extdata", "PBC-variableDetails.csv", package="bllflow"))

# Create a bllFlow R object for the PBC model using the above variables as args
pbcModel <- bllflow::BLLFlow(pbc, variablesSheet, variableDetailsSheet)

# Clean the data
cleanedPbcModel <- bllflow::clean.Min(pbcModel)

# If you wish to be updated in the log on what the function does set print to true
cleanedPbcModel <- bllflow::clean.Min(cleanedPbcModel, print=TRUE)

}
