% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP-plot.R
\name{op_plot}
\alias{op_plot}
\title{Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.}
\usage{
op_plot(data, observed, predicted, title=NA, xlab= NA)
}
\arguments{
\item{data}{a data.frame with observed events and predicted risk estimates}

\item{observed}{variable of observed observation from data.frame, object of class string}

\item{predicted}{variable of predicted observations from data.frame, object of class string}

\item{title}{an overall title for the plot (optional), object of class string}

\item{xlab}{a title for the x axis (optional), object of class string}
}
\description{
Observed vs Predicted Plot
 Creates an observed versus predicted bar chart with scatter plot overlay using plotly. Used for model fit visualization.
}
\author{
Molly Campbell
}
